Dto 정리
