+++ src/main/java/com/tiketeer/Tiketeer/domain/ticketing/controller/dto/PatchTicketingRequestDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.tiketeer.Tiketeer.domain.ticketing.service.dto.UpdateTicketingCommandDto;
+import com.tiketeer.Tiketeer.domain.ticketing.usecase.dto.UpdateTicketingCommandDto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
