+++ src/main/java/com/tiketeer/Tiketeer/domain/ticketing/usecase/dto/CreateTicketingResultDto.java
@@ -1,4 +1,4 @@
-package com.tiketeer.Tiketeer.domain.ticketing.service.dto;
+package com.tiketeer.Tiketeer.domain.ticketing.usecase.dto;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
