+++ src/main/java/com/tiketeer/Tiketeer/domain/ticketing/usecase/dto/DeleteTicketingCommandDto.java
@@ -1,4 +1,4 @@
-package com.tiketeer.Tiketeer.domain.ticket.service.dto;
+package com.tiketeer.Tiketeer.domain.ticketing.usecase.dto;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -9,16 +9,15 @@ import lombok.ToString;
 
 @Getter
 @ToString
-public class DropNumOfTicketsUnderSomeTicketingCommandDto {
+public class DeleteTicketingCommandDto {
 	private final UUID ticketingId;
-	private final int numOfTickets;
+	private final String memberEmail;
 	private LocalDateTime commandCreatedAt = LocalDateTime.now();
 
 	@Builder
-	public DropNumOfTicketsUnderSomeTicketingCommandDto(UUID ticketingId, int numOfTickets,
-		LocalDateTime commandCreatedAt) {
+	public DeleteTicketingCommandDto(UUID ticketingId, String memberEmail, LocalDateTime commandCreatedAt) {
 		this.ticketingId = ticketingId;
-		this.numOfTickets = numOfTickets;
+		this.memberEmail = memberEmail;
 		if (commandCreatedAt != null) {
 			this.commandCreatedAt = commandCreatedAt;
 		}
