+++ src/main/java/com/tiketeer/Tiketeer/domain/ticketing/usecase/dto/GetTicketingResultDto.java
@@ -1,4 +1,4 @@
-package com.tiketeer.Tiketeer.domain.ticketing.service.dto;
+package com.tiketeer.Tiketeer.domain.ticketing.usecase.dto;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -7,52 +7,49 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
-@Getter
 @ToString
-public class UpdateTicketingCommandDto {
+@Getter
+public class GetTicketingResultDto {
 	private final UUID ticketingId;
-	private final String email;
 	private final String title;
 	private final String description;
 	private final String location;
 	private final String category;
 	private final Integer runningMinutes;
 	private final Integer stock;
+	private final Integer remainedStock;
 	private final Long price;
 	private final LocalDateTime eventTime;
 	private final LocalDateTime saleStart;
 	private final LocalDateTime saleEnd;
-	private LocalDateTime commandCreatedAt = LocalDateTime.now();
+	private final LocalDateTime createdAt;
+	private final String owner;
 
 	@Builder
-	public UpdateTicketingCommandDto(
-		UUID ticketingId,
-		String email,
-		String title,
+	public GetTicketingResultDto(UUID ticketingId, String title,
 		String description,
 		String location,
 		String category,
 		Integer runningMinutes,
 		Integer stock,
+		Integer remainedStock,
 		Long price,
 		LocalDateTime eventTime,
 		LocalDateTime saleStart,
-		LocalDateTime saleEnd,
-		LocalDateTime commandCreatedAt) {
+		LocalDateTime saleEnd, LocalDateTime createdAt, String owner) {
 		this.ticketingId = ticketingId;
-		this.email = email;
 		this.title = title;
 		this.description = description;
 		this.location = location;
 		this.category = category;
 		this.runningMinutes = runningMinutes;
 		this.stock = stock;
+		this.remainedStock = remainedStock;
 		this.price = price;
 		this.eventTime = eventTime;
 		this.saleStart = saleStart;
 		this.saleEnd = saleEnd;
-		if (commandCreatedAt != null) {
-			this.commandCreatedAt = commandCreatedAt;
-		}
+		this.createdAt = createdAt;
+		this.owner = owner;
 	}
 }
